(PC-23404)[PRO] refactor: Banner component, split LinkNodes into a single component and move BannerLayoutinto Banner component
